backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
pydantic==2.11.7
faster-whisper==1.1.0
pyannote.audio==3.3.2
torch==2.8.0
requests==2.32.5
